backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
requests>=2.31.0
pathlib
ffmpeg-python
lmstudio
huggingface-hub
transformers
torchaudio
torch